[1586] Extract CoursesEditorCLI which is solely responsible for asking for user input
